requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2025.1
django-filter==25.1
colorama==0.4.6
iniconfig==2.0.0
packaging==24.2
packaging==24.2
pytest==8.3.4
pytest-django==4.10.0